Move compression type deduction to the factory decompressor
