+++ src/shared_modules/content_manager/tests/unit/factoryDownloader_test.cpp
@@ -53,93 +53,18 @@ TEST_F(FactoryDownloaderTest, CreateS3Downloader)
 }
 
 /**
- * @brief Check the creation of a OfflineDownloader for downloading a raw file.
+ * @brief Check the creation of a OfflineDownloader.
  *
  */
-TEST_F(FactoryDownloaderTest, CreateOfflineDownloaderRawFile)
+TEST_F(FactoryDownloaderTest, CreateOfflineDownloader)
 {
-    auto config = R"(
-        {
-            "contentSource": "offline",
-            "url": "file:///home/user/file.txt",
-            "compressionType": "ignored"
-        }
-    )"_json;
-
-    const auto expectedConfig = R"(
-        {
-            "contentSource": "offline",
-            "url": "file:///home/user/file.txt",
-            "compressionType": "raw"
-        }
-    )"_json;
+    auto config = R"({"contentSource":"offline"})"_json;
 
     // Create the downloader.
     std::shared_ptr<AbstractHandler<std::shared_ptr<UpdaterContext>>> spDownloader {};
-
     EXPECT_NO_THROW(spDownloader = FactoryDownloader::create(config));
+
     EXPECT_TRUE(std::dynamic_pointer_cast<OfflineDownloader>(spDownloader));
-    EXPECT_EQ(config, expectedConfig);
-}
-
-/**
- * @brief Check the creation of a OfflineDownloader for downloading a compressed file.
- *
- */
-TEST_F(FactoryDownloaderTest, CreateOfflineDownloaderCompressedFile)
-{
-    auto config = R"(
-        {
-            "contentSource": "offline",
-            "url": "file:///home/user/file.txt.gz",
-            "compressionType": "ignored"
-        }
-    )"_json;
-
-    const auto expectedConfig = R"(
-        {
-            "contentSource": "offline",
-            "url": "file:///home/user/file.txt.gz",
-            "compressionType": "gzip"
-        }
-    )"_json;
-
-    // Create the downloader.
-    std::shared_ptr<AbstractHandler<std::shared_ptr<UpdaterContext>>> spDownloader {};
-
-    EXPECT_NO_THROW(spDownloader = FactoryDownloader::create(config));
-    EXPECT_TRUE(std::dynamic_pointer_cast<OfflineDownloader>(spDownloader));
-    EXPECT_EQ(config, expectedConfig);
-}
-
-/**
- * @brief Check the creation of a OfflineDownloader for downloading a file with no extension.
- *
- */
-TEST_F(FactoryDownloaderTest, CreateOfflineDownloaderNoExtension)
-{
-    auto config = R"(
-        {
-            "contentSource": "offline",
-            "url": "file:///home/user/file_without_extension",
-            "compressionType": "ignored"
-        }
-    )"_json;
-
-    const auto expectedConfig = R"(
-        {
-            "contentSource": "offline",
-            "url": "file:///home/user/file_without_extension",
-            "compressionType": "raw"
-        }
-    )"_json;
-
-    // Create the downloader.
-    std::shared_ptr<AbstractHandler<std::shared_ptr<UpdaterContext>>> spDownloader {};
-
-    EXPECT_NO_THROW(spDownloader = FactoryDownloader::create(config));
-    EXPECT_TRUE(std::dynamic_pointer_cast<OfflineDownloader>(spDownloader));
-    EXPECT_EQ(config, expectedConfig);
 }
 
 /*
